further changes

+++ packages/trpc/server/routers/viewer/teams/managedEvents/getManagedEventUsersToReassign.schema.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 
 export const ZGetManagedEventUsersToReassignInputSchema = z.object({
-  bookingId: z.number(),
-  cursor: z.number().optional(), // For pagination
+  bookingId: z.number().int(),
+  cursor: z.number().int().optional(), // For pagination
   limit: z.number().min(1).max(100).optional(),
   searchTerm: z.string().optional(),
 });
